Add `-nofullscreen` to the command line
This has no real effect on SDL since we run with the `dummy` driver that
just does everything offscreen. However, Chocolate Doom by default
pauses for a second on startup if running fullscreen, so that the
display has time to settle if the monitor changes modes. Since statcheck
runs through hundreds of fairly short demos, this equates to a useless
pause that makes the whole run take much longer.